MAINT: Correct typing issues
