+++ pandas-stubs/_libs/lib.pyi
@@ -1,8 +1,3 @@
-from enum import Enum
-
 no_default = None
-
-class NoDefault(Enum):
-    no_default: int
 
 def infer_dtype(value: object, skipna: bool = ...) -> str: ...
