+++ pandas-stubs/io/clipboards.pyi
@@ -15,6 +15,7 @@ from pandas.core.series import Series
 from pandas._typing import (
     CompressionOptions,
     CSVEngine,
+    CSVQuoting,
     DtypeArg,
     StorageOptions,
     npt,
@@ -71,7 +72,7 @@ def read_clipboard(
     decimal: str = ...,
     lineterminator: str | None = ...,
     quotechar: str = ...,
-    quoting: int = ...,
+    quoting: CSVQuoting | None = ...,
     doublequote: bool = ...,
     escapechar: str | None = ...,
     comment: str | None = ...,
@@ -137,7 +138,7 @@ def read_clipboard(
     decimal: str = ...,
     lineterminator: str | None = ...,
     quotechar: str = ...,
-    quoting: int = ...,
+    quoting: CSVQuoting | None = ...,
     doublequote: bool = ...,
     escapechar: str | None = ...,
     comment: str | None = ...,
@@ -203,7 +204,7 @@ def read_clipboard(
     decimal: str = ...,
     lineterminator: str | None = ...,
     quotechar: str = ...,
-    quoting: int = ...,
+    quoting: CSVQuoting | None = ...,
     doublequote: bool = ...,
     escapechar: str | None = ...,
     comment: str | None = ...,
