+++ tests/test_io.py
@@ -7,6 +7,7 @@ from typing import (
     Union,
 )
 
+import numpy as np
 from packaging.version import parse
 import pandas as pd
 from pandas import (
@@ -365,6 +366,48 @@ def test_read_csv_iterator():
             assert_type(read_csv(pathlib.Path(path), chunksize=1), TextFileReader),
             TextFileReader,
         )
+
+
+def test_types_read_csv() -> None:
+    df = pd.DataFrame(data={"col1": [1, 2], "col2": [3, 4]})
+    csv_df: str = df.to_csv()
+
+    with ensure_clean() as path:
+        df.to_csv(path)
+        df2: pd.DataFrame = pd.read_csv(path)
+        df3: pd.DataFrame = pd.read_csv(path, sep="a")
+        df4: pd.DataFrame = pd.read_csv(
+            path,
+            header=None,
+        )
+        df5: pd.DataFrame = pd.read_csv(
+            path, engine="python", true_values=["no", "No", "NO"], na_filter=False
+        )
+        df6: pd.DataFrame = pd.read_csv(
+            path,
+            skiprows=lambda x: x in [0, 2],
+            skip_blank_lines=True,
+            dayfirst=False,
+        )
+        df7: pd.DataFrame = pd.read_csv(path, nrows=2)
+        df8: pd.DataFrame = pd.read_csv(path, dtype={"a": float, "b": int})
+        df9: pd.DataFrame = pd.read_csv(path, usecols=["col1"])
+        df10: pd.DataFrame = pd.read_csv(path, usecols=[0])
+        df11: pd.DataFrame = pd.read_csv(path, usecols=np.array([0]))
+        df12: pd.DataFrame = pd.read_csv(path, usecols=("col1",))
+        df13: pd.DataFrame = pd.read_csv(path, usecols=pd.Series(data=["col1"]))
+
+        tfr1: TextFileReader = pd.read_csv(path, nrows=2, iterator=True, chunksize=3)
+        tfr1.close()
+
+        tfr2: TextFileReader = pd.read_csv(path, nrows=2, chunksize=1)
+        tfr2.close()
+
+        tfr3: TextFileReader = pd.read_csv(path, nrows=2, iterator=False, chunksize=1)
+        tfr3.close()
+
+        tfr4: TextFileReader = pd.read_csv(path, nrows=2, iterator=True)
+        tfr4.close()
 
 
 def test_read_table():
